virtio: Fix type cast from integer value to pointer for Windows
This Linux header is included when building hw/char/virtio-console.o,
hw/virtio/virtio-rng.o, hw/virtio/virtio-pci.o, hw/virtio/virtio-bus.o
and hw/virtio/virtio-mmio.o for Windows. On 64 bit Windows, type casts
from unsigned long to pointer cause a compiler warning.

More fixes might be needed for other unsigned long values, for example
parameter 'align'.

+++ include/standard-headers/linux/virtio_ring.h
@@ -143,7 +143,7 @@ static inline void vring_init(struct vring *vr, unsigned int num, void *p,
 	vr->num = num;
 	vr->desc = p;
 	vr->avail = p + num*sizeof(struct vring_desc);
-	vr->used = (void *)(((unsigned long)&vr->avail->ring[num] + sizeof(__virtio16)
+	vr->used = (void *)(((uintptr_t)&vr->avail->ring[num] + sizeof(__virtio16)
 		+ align-1) & ~(align - 1));
 }
